fix(runner): le verrou de campagne ne s'exécutait pas, le sondage ne produisait rien
`pg_try_advisory_xact_lock(%s, f.id)` sur une colonne BIGSERIAL n'a pas de
signature en Postgres : `(int, bigint)` n'existe pas. La requête levait
`UndefinedFunction` à chaque appel, le `try` de `_produire_pour_une_campagne`
l'avalait, et le sondage rendait « aucun travail » — depuis la mise en service,
sans avoir produit une seule fois.

Mesuré le 07/09/2026 en activant la garde sur le canari : trois workers
sondaient en production depuis des jours, `/opt/oto-runner/passages` était vide,
aucun travail n'était jamais passé par la file. Le défaut était invisible parce
que le repli de l'appelant transforme « je n'ai pas pu regarder » en « il n'y a
rien à faire », deux réponses que le client ne distingue pas.

Aucun banc ne pouvait le voir : tous mockent `db.campagne_a_servir` et prouvent
ce que la capacité fait du résultat, jamais que la requête s'exécute.
`tests/test_campagne_a_servir_db.py` l'exerce en SQL réel — trois de ses quatre
tests tombent quand on retire le cast, le quatrième n'utilisant pas ce verrou.

Changelog: fixed

+++ oto_mcp/db/runner_fleets.py
@@ -138,7 +138,14 @@ def campagne_a_servir(org_id: int) -> Optional[dict]:
             SELECT {_COLS}
               FROM runner_fleets f
              WHERE f.org_id = %s
-               AND pg_try_advisory_xact_lock(%s, f.id)
+               -- `f.id::int` et non `f.id` : la colonne est BIGSERIAL, et
+               -- Postgres n'offre que `(bigint)` ou `(int, int)` — jamais
+               -- `(int, bigint)`. Sans le cast, la requête LÈVE
+               -- `UndefinedFunction`, le `try` de l'appelant l'avale, et le
+               -- sondage rend « aucun travail » pour toujours : mesuré le
+               -- 07/09/2026 sur le canari, la production de travail n'avait
+               -- jamais pu s'exécuter une seule fois.
+               AND pg_try_advisory_xact_lock(%s, f.id::int)
                AND f.status IN ('armed', 'running')
                AND NOT EXISTS (SELECT 1 FROM runner_jobs j
                                 WHERE j.fleet_id = f.id AND j.status = 'pending')
